Remove thead-based tests and improve test_messages.py
